pokemon_data_scraper/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clickhouse-driver==0.2.3
coloredlogs==15.0.1
cryptography==36.0.2
Cython==0.29.28
h11==0.13.0
humanfriendly==10.0
idna==3.3
joblib==1.1.0
numpy==1.22.3
outcome==1.1.0
pandas==1.4.2
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
timestring==1.6.4
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wsproto==1.1.0